Better modularize compiledb implementation
