+++ compiledb/utils.py
@@ -18,12 +18,21 @@
 #   You should have received a copy of the GNU General Public License
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
+from __future__ import print_function
 
 import sys
 
-from compiledb.generator import generate
 
-if(__name__ == "__main__"):
-    sys.exit(generate())
+def input_file(path):
+    return sys.stdin if path is None else open(path, "r")
+
+
+def output_file(path):
+    return sys.stdout if path is None else open(path, "w")
+
+
+def msg(*args, **kwargs):
+    print(*args, file=sys.stderr, **kwargs)
 
 # ex: ts=2 sw=4 et filetype=python
+
